api_server/requirements.txt
dj-rest-auth==2.2.4
Django==4.0.6
djangorestframework==3.13.1
ginza==5.1.1
ja-ginza-electra==5.1.0
spacy==3.2.4